[REF] r_allocation: split allocation file
