+++ resource_planning/models/__init__.py
@@ -1,3 +1,4 @@
 from . import resource_location
 from . import resource
+from . import resource_allocation
 from . import res_config
